moving installation script to Nextcloud Pod
